adds rating to UI

+++ HikingApp/app/src/androidTest/java/ch/epfl/sweng/team7/mockServer/MockServer.java
@@ -280,7 +280,7 @@ public class MockServer implements DatabaseClient {
      * @param hikeId the hike the vote is about
      * @param value the vote value
      */
-    public void postVote(long hikeId, int value) throws DatabaseClientException {
+    public void postVote(long hikeId, float value) throws DatabaseClientException {
         throw new DatabaseClientException("Not implemented.");
     }
